requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
Pillow>=10.4.0
requests>=2.32.0
openpyxl>=3.1.5
mcp
Faker
matplotlib>=3.7.0
numpy>=1.24.0